git-svn-id: https://root.cern.ch/svn/root/branches/dev/roostats@33495 27541ba8-7e3a-0410-8455-c3a389f83636

+++ gui/gui/src/TGFileBrowser.cxx
@@ -218,8 +218,9 @@ TGFileBrowser::~TGFileBrowser()
    delete fContextMenu;
    delete fListTree;
    if (fRootIcon) fClient->FreePicture(fRootIcon);
+   if (fCachedPic && (fCachedPic != fFileIcon))
+      fClient->FreePicture(fCachedPic);
    if (fFileIcon) fClient->FreePicture(fFileIcon);
-   if (fCachedPic) fClient->FreePicture(fCachedPic);
    Cleanup();
 }
 
@@ -259,7 +260,7 @@ void TGFileBrowser::Add(TObject *obj, const char *name, Int_t check)
                                                       pic, pic, kTRUE);
             if ((pic != fFileIcon) && (pic != fCachedPic))
                fClient->FreePicture(pic);
-            fListTree->CheckItem(item, (Bool_t)check);
+            if (item) fListTree->CheckItem(item, (Bool_t)check);
             TString tip(obj->ClassName());
             if (obj->GetTitle()) {
                tip += " ";
@@ -304,7 +305,8 @@ void TGFileBrowser::Add(TObject *obj, const char *name, Int_t check)
                TGListTreeItem *item = fListTree->AddItem(fListLevel, name, obj, pic, pic);
                if ((pic != fFileIcon) && (pic != fCachedPic))
                   fClient->FreePicture(pic);
-               item->SetDNDSource(kTRUE);
+               if (item && obj && obj->InheritsFrom("TObject"))
+                  item->SetDNDSource(kTRUE);
             }
          }
       }
@@ -387,7 +389,7 @@ void TGFileBrowser::BrowseObj(TObject *obj)
          if (named)
             curdrive = named->GetName();
          else
-            curdrive = StrDup("C:");
+            curdrive = "C:";
          TIter next(volumes);
          TNamed *drive;
          while ((drive = (TNamed *)next())) {
@@ -547,8 +549,9 @@ void TGFileBrowser::Update()
          fListTree->GetPathnameFromItem(curr->GetParent(), path);
          if (strlen(path) > 1) {
             TString dirpath = FullPathName(curr->GetParent());
-            gSystem->GetPathInfo(dirpath.Data(), &id, &size, &flags, &modtime);
-            if (flags & 2) {
+            Int_t res = gSystem->GetPathInfo(dirpath.Data(), &id, &size, 
+                                             &flags, &modtime);
+            if ((res == 0) && (flags & 2)) {
                TString fullpath = FullPathName(curr);
                if (gSystem->AccessPathName(fullpath.Data()))
                   fListTree->DeleteItem(curr);
@@ -1213,6 +1216,7 @@ void TGFileBrowser::GetObjPicture(const TGPicture **pic, TObject *obj)
    // Retrieve icons associated with class "name". Association is made
    // via the user's ~/.root.mimes file or via $ROOTSYS/etc/root.mimes.
 
+   const char *clname = 0;
    TClass *objClass = 0;
    static TImage *im = 0;
    if (!im) {
@@ -1221,28 +1225,32 @@ void TGFileBrowser::GetObjPicture(const TGPicture **pic, TObject *obj)
 
    if (obj->IsA() == TClass::Class()) {
       objClass = obj->IsA();
+      if (objClass) 
+         clname = objClass->GetName();
    }
    else if (obj->InheritsFrom("TKey")) {
-      const char *clname = (const char *)gROOT->ProcessLine(TString::Format("((TKey *)0x%lx)->GetClassName();", obj));
-      if (clname)
-         objClass = TClass::GetClass(clname);
+      clname = (char *)gROOT->ProcessLine(TString::Format("((TKey *)0x%lx)->GetClassName();", obj));
    }
    else if (obj->InheritsFrom("TKeyMapFile")) {
-      const char *title = (const char *)gROOT->ProcessLine(TString::Format("((TKeyMapFile *)0x%lx)->GetTitle();", obj));
-      if (title)
-         objClass = TClass::GetClass(title);
+      clname = (char *)gROOT->ProcessLine(TString::Format("((TKeyMapFile *)0x%lx)->GetTitle();", obj));
    }
    else if (obj->InheritsFrom("TRemoteObject")) {
       // special case for remote object: get real object class
       TRemoteObject *robj = (TRemoteObject *)obj;
       if (!strcmp(robj->GetClassName(), "TKey"))
-         objClass = TClass::GetClass(robj->GetKeyClassName());
+         clname = robj->GetKeyClassName();
       else
-         objClass = TClass::GetClass(robj->GetClassName());
+         clname = robj->GetClassName();
    }
-   else
+   else {
       objClass = obj->IsA();
-   const char *name = obj->GetIconName() ? obj->GetIconName() : objClass->GetName();
+      if (objClass) 
+         clname = objClass->GetName();
+   }
+   if (!clname) { 
+      clname = "Unknown";
+   }
+   const char *name = obj->GetIconName() ? obj->GetIconName() : clname;
    TString xpm_magic(name, 3);
    Bool_t xpm = xpm_magic == "/* ";
    const char *iconname = xpm ? obj->GetName() : name;
